little restructure
